+++ Sphinx/ChangeLog.rst
@@ -2,6 +2,7 @@ ChangeLog
 =========
 * October 7, 2014
 
+  - Corrected local filename bug in ``_load_media()``. (Thanks to Sergey Sorokin.)
   - Updated documentation.
 
 * 01.08.00 --- October 4, 2014